+++ CHANGELOG.md
@@ -85,3 +85,4 @@ Our Keycloak version is working well with PostgreSQL database. For using other S
 - Being possible to accept terms and conditions before User is saved in Keycloak during first broker login. Follow GDPR.
 - Being possible to add realm default scopes during Dynamic Client Registration/ OpenID Federation when scopes are including in client representation
 - Problem saving OpenId Federation disabled [RCIAM-636](https://tts.grnet.gr/jira/browse/RCIAM-636)
+- Access token type based on related user session[RCIAM-628](https://tts.grnet.gr/jira/browse/RCIAM-628)